fix(frontend): Automatically choose client/server-side API client
