+++ autogpt_platform/frontend/src/lib/autogpt-server-api/index.ts
@@ -1,5 +1,6 @@
 import { AutoGPTServerAPI } from "./client";
 
+// TODO: rename to BackendAPI
 export default AutoGPTServerAPI;
 export * from "./client";
 export * from "./context";
